Implement dir opening and (auto) saving
A button has been added to allow the user to pick a directory to load images from.

Auto saving in the form of logging the undo and redo stacks to a json file in the image directory has been implemented. This file is updated every time an change to the stack is made, resulting in a lot of reads and writes to the json file. This could become untenable for large datasets, so this may need to be changed in the future.