Fix detrimental history writes and add an export button.
Originally, upon the opening of a new image the history would automatically be written to. This became a problem after calling reset_canvas(), as this would completely wipe the history. This write is no longer called.

The export button will export the current bounding boxes and labels as a json file to a chosen directory. It currently creates class called Exporter and pops up a dummy window - I'm not sure whether I will keep the pop up or not.