Added status bar and marginally better history management
The status bar shows the most recent action on the left and various pieces of information about the current file on the right. I might want to add more file information later.

The history is now managed through a class called HistoryManager. This is then used to write to the .json file after most changes. I think the history should only be written to on file openings and upon quitting the application.