Change to use conventional format for storage / exporting
Though the system I had implemented beforehand worked well, in order for the program to be at all useful - it has to conform to a conventional standards. I have started implementing the COCO (Common Objects in Context) format and intend to implement the equivalent for Pascal VOC too.

The issue is that in implementing this, it has superseded lots of legacy code that I and I haven't removed it all yet.

This also introduces a need for an InfoWidget, for the user to input information that cannot be immediately gleaned from the image and annotations themselves.