Find date_captured of each image based on its metadata
This uses the exif data of the image to get the date that the image was taken, so as to comply with coco standards. If the exif data cannot be found, data_captured is set as "today's date".

Date_captured is an optional field in the coco format, so this function will probably be optional in the future.