New Annotation class and (basic) undo functionality
After examining the COCO format more closely, I realised that each annotation has one bounding box and one or more segmentation polygons. This meant that I had to implement an Annotation class to account for this, which in turn broke the undo and redo stack. I have the undo stack working again.