Prevent undo/redo stack desync on invalid segment
Previously the prevention of an invalid segment being drawn worked, but that segment object was still appended to the undo/redo stack when it shouldn't have been. This fixes that issue.

This also adds a new file called operations.py within which the intersection calculations are stored (and within which the RLE algorithm will be stored)